[#noissue] Extract ApplicationMapView
