+++ web/src/main/java/com/navercorp/pinpoint/web/applicationmap/ApplicationMap.java
@@ -15,8 +15,6 @@
 
 package com.navercorp.pinpoint.web.applicationmap;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.navercorp.pinpoint.common.timeseries.time.Range;
 import com.navercorp.pinpoint.web.applicationmap.link.Link;
 import com.navercorp.pinpoint.web.applicationmap.nodes.Node;
@@ -28,13 +26,9 @@ import java.util.Collection;
  */
 public interface ApplicationMap {
 
-
-    @JsonProperty("nodeDataArray")
     Collection<Node> getNodes();
 
-    @JsonProperty("linkDataArray")
     Collection<Link> getLinks();
 
-    @JsonInclude(JsonInclude.Include.NON_NULL)
     Range getRange();
 }
